[ISSUE #6301] add unit tests.
